+++ pyproject.toml
@@ -1,0 +1,31 @@
+[tool.poetry]
+name = "dumbrl"
+version = "0.1.0"
+description = "A collection of simple RL environments for experimentation."
+authors = ["Generic <you@example.com>"]
+readme = "README.md"
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+gymnasium = "^0.29.1"
+numpy = "^1.26.3"
+stable-baselines3 = "^2.2.1"
+rich = "^13.7.0"
+
+[tool.poetry.dev-dependencies]
+black = "^22.3.0"  # Ensure you have the latest version
+pre-commit = "^2.17.0"  # Ensure you have the latest version
+
+[tool.poetry.group.dev.dependencies]
+isort = "^5.13.2"
+
+[tool.black]
+line-length = 80
+
+[tool.isort]
+profile = "black"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
